sova-tts-engine/requirements.txt
# torch===1.4.0 https://download.pytorch.org/whl/torch_stable.html
scipy==1.4.1
numpy==1.18.5
librosa==0.7.2
matplotlib==2.1.0
pillow
torch_optimizer==0.0.1a15
-e git+https://github.com/sovaai/sova-tts-tps#egg=TPS
